VinFast-SOA-Project-master/api-gateway/requirements.txt
flask
flask-cors
requests
pyjwt
flask-socketio
eventlet
python-socketio